fix: harden APIs per TIME-27.1 audit findings
- convert: Galileo/BeiDou into_scale_with_checked now route through the
  GPS detector and report AmbiguousLeapSecond inside the insertion
  window (previously always Exact)
- leap: RuntimeLeapSeconds::from_slice validates entries via try_extend;
  add LeapExtendError::EmptyTable and validated LeapSeconds::try_from_slice;
  fix "tham"/BUILTIN_YABLE doc typos
- duration: from_* constructors panic deterministically on overflow in all
  build profiles; add checked_from_minutes/hours/days
- time: from_week_tow takes week as u32 (BREAKING); document to_civil
  infallibility invariant; fix as_parts f64 doc claim
- benches: "empty, fallback" case uses static LeapSeconds::from_slice(&[])
  since empty runtime tables are now rejected; README tables renamed

+++ examples/embedded_minimal.rs
@@ -20,7 +20,7 @@ fn panic(_info: &core::panic::PanicInfo) -> ! {
 
 #[inline(never)]
 fn probe_from_week_tow(
-    week: u16,
+    week: u32,
     tow: DurationParts,
 ) -> Time<Gps> {
     Time::<Gps>::from_week_tow(week, tow).unwrap()
